+++ R/package.R
@@ -16,8 +16,12 @@
 #'
 #' @import rmarkdown
 #' @import htmltools
+#' @import sass
+#' @import bslib
 #' @importFrom jsonlite toJSON
 #' @importFrom tools file_path_sans_ext
+#' @importFrom utils packageVersion
+#' @importFrom grDevices col2rgb
 #'
 #' @docType package
 #' @name flexdashboard